RAG/Pipelines/requirements.txt
PyPDF2
beautifulsoup4
requests
spacy
pdfplumber
python-docx
PyMuPDF
selenium
pandas
sentence-transformers
scikit-learn
numpy

#chunking
openai
sentence-transformers
boto3
cohere
scikit-learn
numpy


# vector database
faiss-cpu
chromadb
pinecone-client
weaviate-client


# app UI
fastapi
uvicorn
streamlit
# 